+++ packages/spotlight/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact spotlight module, newest changes on the top.
 
+## [unreleased]
+
+### Removed
+
+- `@enact/spotlight.spotlightDirections` export and replaced it with `getDirection()` 
+
 ## [1.0.0-beta.1] - 2016-12-30
 
 ### Added